+++ ureport/public/tests.py
@@ -46,12 +46,12 @@ class PublicTest(UreportTest):
         self.login(self.admin)
         response = self.client.get(edit_url, SERVER_NAME="nigeria.ureport.io")
         self.assertTrue("form" in response.context)
-        self.assertEqual(len(response.context["form"].fields), 42)
+        self.assertEqual(len(response.context["form"].fields), 43)
 
         self.login(self.superuser)
         response = self.client.get(edit_url, SERVER_NAME="nigeria.ureport.io")
         self.assertTrue("form" in response.context)
-        self.assertEqual(len(response.context["form"].fields), 66)
+        self.assertEqual(len(response.context["form"].fields), 67)
 
     def test_count(self):
         count_url = reverse("public.count")